Merge into Master

See merge request 2021ss-sepm-pr-group/21ss-sepm-pr-inso-08!33

+++ frontend/src/app/services/auth.service.ts
@@ -1,48 +1,54 @@
-import {Injectable} from '@angular/core';
-import {AuthRequest} from '../dtos/auth-request';
-import {Observable} from 'rxjs';
-import {HttpClient} from '@angular/common/http';
-import {tap} from 'rxjs/operators';
+import { Injectable } from '@angular/core';
+import { AuthRequest } from '../dtos/auth-request';
+import { Observable } from 'rxjs';
+import { HttpClient } from '@angular/common/http';
+import { tap } from 'rxjs/operators';
 import jwt_decode from 'jwt-decode';
-import {Globals} from '../global/globals';
+import { Globals } from '../global/globals';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class AuthService {
-
   private authBaseUri: string = this.globals.backendUri + '/authentication';
 
-  constructor(private httpClient: HttpClient, private globals: Globals) {
-  }
+  constructor(private httpClient: HttpClient, private globals: Globals) {}
 
   /**
    * Login in the user. If it was successful, a valid JWT token will be stored
    *
    * @param authRequest User data
    */
-  loginUser(authRequest: AuthRequest): Observable<string> {
-    return this.httpClient.post(this.authBaseUri, authRequest, {responseType: 'text'})
-      .pipe(
-        tap((authResponse: string) => this.setToken(authResponse))
-      );
+  loginUser(authRequest: AuthRequest, keepLogin: boolean): Observable<string> {
+    console.log(keepLogin);
+    return this.httpClient
+      .post(this.authBaseUri, authRequest, { responseType: 'text' })
+      .pipe(tap((authResponse: string) => this.setToken(authResponse, keepLogin)));
   }
-
 
   /**
    * Check if a valid JWT token is saved in the localStorage
    */
-  isLoggedIn() {
-    return !!this.getToken() && (this.getTokenExpirationDate(this.getToken()).valueOf() > new Date().valueOf());
+  isLoggedIn(): boolean {
+    return (
+      !!this.getToken() &&
+      this.getTokenExpirationDate(this.getToken()).valueOf() >
+        new Date().valueOf()
+    );
   }
 
-  logoutUser() {
+  logoutUser(): void {
     console.log('Logout');
     localStorage.removeItem('authToken');
+    sessionStorage.removeItem('authToken');
   }
 
-  getToken() {
-    return localStorage.getItem('authToken');
+  getToken(): string {
+    let token = localStorage.getItem('authToken');
+    if (token == null) {
+      token = sessionStorage.getItem('authToken');
+    }
+    return token;
   }
 
   /**
@@ -66,7 +72,7 @@ export class AuthService {
   /**
    * Returns the email of the user based on the current token
    */
-   getUserEmail() {
+  getUserEmail() {
     if (this.getToken() != null) {
       const decoded: any = jwt_decode(this.getToken());
       return decoded.sub;
@@ -74,12 +80,15 @@ export class AuthService {
     return null;
   }
 
-  private setToken(authResponse: string) {
-    localStorage.setItem('authToken', authResponse);
+  private setToken(authResponse: string, keepLogin: boolean) {
+    if (keepLogin) {
+      localStorage.setItem('authToken', authResponse);
+    } else {
+      sessionStorage.setItem('authToken', authResponse);
+    }
   }
 
   private getTokenExpirationDate(token: string): Date {
-
     const decoded: any = jwt_decode(token);
     if (decoded.exp === undefined) {
       return null;
@@ -89,5 +98,4 @@ export class AuthService {
     date.setUTCSeconds(decoded.exp);
     return date;
   }
-
 }
